Added a simple example/dev setup
A simple example of Morph in action. This also acts as a dev environment that allows you to both tweak the examples as well as the main morph-forms library itself. We're making use of the brilliant react hot-loader so any changes are pushed to the browser. This is a nice easy way for people to play around with, and get to grips with Morph.