Check if field.data is undefined rather than falsy
We don't mind if field.data is null (which it will be on component mount unless a different default value is provided). If field.data is null we want that to go into the values tree as null rather than returning the entire data object which it was doing previously.