Add basic validation and error handling
This introduces a set of simple (but powerful!) validators which can be passed to fields. These will be run every time the value of the field changes and as the change propagates up the form the validity of higher order components (HOC) will be checked as well. This allows us to set validators which check if two fields in a fieldset are equal, or even if two fieldsets are equal.

Errors are held in the state of each field and exposed through the container. Errors are not bundled together into the state of HOCs as a HOC can generate it's own errors which we want to keep separate from any field errors (e.g. adding a validator to a fieldset to check that two fields within the fieldset are equal).

The validity does propagate though, so if a field is invalid this will propagate up the form making the form itself invalid.